several improvements better targeting system and placing towers +menu, abstract class and interface( for future towers and enemies)

+++ src/Helpers/Clock.java
@@ -41,7 +41,7 @@ public class Clock {
         totalTime += deltaTime;
     }
 
-    public static void changeMultiplier(int change) {
+    public static void changeMultiplier(float change) {
         if (!(multiplier + change < -1 && multiplier + change > 7)) {
             multiplier += change;
         }
